# backend_gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrepSense Gateway API",
...
    "version": "1.4.0",
    "contact": {
      "name": "PrepSense Development Team",
      "email": "dev@prepsense.com",
      "url": "https://github.com/dankimjw/PrepSense"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8002",
      "description": "Testing server"
    },
    {
      "url": "https://api.prepsense.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/recipe-consumption/cook": {
      "post": {
        "tags": [
          "recipes",
          "Recipe Consumption"
        ],
        "summary": "Cook a recipe and update pantry",
        "description": "Cook a recipe and subtract ingredients from pantry\n\nThis endpoint:\n1. Validates that the user has sufficient ingredients\n2. Subtracts the used quantities from pantry items\n3. Records the cooking event\n4. Returns updated pantry state",
        "operationId": "cook_recipe_api_v1_recipe_consumption_cook_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Cook Recipe Api V1 Recipe Consumption Cook Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recipe-consumption/check-ingredients": {
      "post": {
        "tags": [
          "recipes",
          "Recipe Consumption"
        ],
        "summary": "Check if user has sufficient ingredients",
        "description": "Check if user has sufficient ingredients to cook a recipe\n\nReturns data in format expected by QuickCompleteModal frontend",
        "operationId": "check_ingredients_availability_api_v1_recipe_consumption_check_ingredients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Check Ingredients Availability Api V1 Recipe Consumption Check Ingredients Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recipe-consumption/quick-complete": {
      "post": {
        "tags": [
          "recipes",
          "Recipe Consumption"
        ],
        "summary": "Quick complete a recipe",
        "description": "Quick complete a recipe by updating pantry quantities based on ingredient selections",
        "operationId": "quick_complete_recipe_api_v1_recipe_consumption_quick_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Quick Complete Recipe Api V1 Recipe Consumption Quick Complete Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CookRecipeRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id",
            "description": "Spoonacular recipe ID"
          },
          "recipe_title": {
            "type": "string",
            "title": "Recipe Title",
            "description": "Recipe title"
          },
          "servings": {
            "type": "integer",
            "title": "Servings",
            "description": "Number of servings being made",
            "default": 1
          },
          "ingredients_used": {
            "items": {
              "$ref": "#/components/schemas/IngredientUsage"
            },
            "type": "array",
            "title": "Ingredients Used",
            "description": "List of ingredients used from pantry"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "recipe_id",
          "recipe_title",
          "ingredients_used"
        ],
        "title": "CookRecipeRequest",
        "description": "Request model for cooking a recipe"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngredientCheckRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id",
            "description": "Spoonacular recipe ID"
          },
          "servings": {
            "type": "integer",
            "title": "Servings",
            "description": "Number of servings being made",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "recipe_id"
        ],
        "title": "IngredientCheckRequest",
        "description": "Request model for checking ingredient availability"
      },
      "IngredientSelection": {
        "properties": {
          "ingredient_name": {
            "type": "string",
            "title": "Ingredient Name",
            "description": "Name of the ingredient"
          },
          "pantry_item_id": {
            "type": "integer",
            "title": "Pantry Item Id",
            "description": "Selected pantry item ID"
          },
          "quantity_to_use": {
            "type": "number",
            "title": "Quantity To Use",
            "description": "Quantity to use"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Unit of measurement"
          }
        },
        "type": "object",
        "required": [
          "ingredient_name",
          "pantry_item_id",
          "quantity_to_use",
          "unit"
        ],
        "title": "IngredientSelection",
        "description": "Model for ingredient selection in quick complete"
      },
      "IngredientUsage": {
        "properties": {
          "pantry_item_id": {
            "type": "integer",
            "title": "Pantry Item Id",
            "description": "ID of the pantry item to use"
          },
          "ingredient_name": {
            "type": "string",
            "title": "Ingredient Name",
            "description": "Name of the ingredient"
          },
          "quantity_used": {
            "type": "number",
            "title": "Quantity Used",
            "description": "Amount of ingredient used"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Unit of measurement"
          }
        },
        "type": "object",
        "required": [
          "pantry_item_id",
          "ingredient_name",
          "quantity_used",
          "unit"
        ],
        "title": "IngredientUsage",
        "description": "Model for ingredient usage in a recipe"
      },
...
      "QuickCompleteRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID"
          },
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id",
            "description": "Spoonacular recipe ID"
          },
          "servings": {
            "type": "integer",
            "title": "Servings",
            "description": "Number of servings being made",
            "default": 1
          },
          "ingredient_selections": {
            "items": {
              "$ref": "#/components/schemas/IngredientSelection"
            },
            "type": "array",
            "title": "Ingredient Selections",
            "description": "Selected ingredients"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "recipe_id",
          "ingredient_selections"
        ],
        "title": "QuickCompleteRequest",
        "description": "Request model for quick completing a recipe"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for external services"
      }
    },
...
  },
...
}